feat(dropdown): add support for hover trigger
