+++ lib/components/popover.vue
@@ -39,5 +39,4 @@
             }
         }
     };
-
 </script>
