+++ lib/mixins/dropdown.js
@@ -1,4 +1,5 @@
 import clickOutMixin from './clickout';
+import triggersMixin from './triggers';
 
 // Determine if an HTML element is visible - Faster than CSS check
 function isVisible(el) {
@@ -21,6 +22,7 @@ const HEADER_SELECTOR = '.dropdown-header';
 const ALL_SELECTOR = [ITEM_SELECTOR, HEADER_SELECTOR].join(',');
 
 export default {
+    mixins: [clickOutMixin, triggersMixin],
     props: {
         id: {
             type: String
@@ -40,109 +42,103 @@ export default {
         right: {
             type: Boolean,
             default: false
+        },
+        debounce: {
+            default: 0
+        },
+        delay: {
+            default: () => { return { show: 0, hide: 150 }}
         }
     },
     data() {
         return {
-            visible: false
+            triggerName: 'dropdown'
         };
     },
     created() {
+        this.triggersCreated();
+
         const listener = el => {
             if (el !== this) {
-                this.visible = false;
+                this.toggle(false);
             }
         };
 
         // To keep one dropdown opened on page
-        this.$root.$on('shown::dropdown', listener);
+        // this.$root.$on('shown::dropdown', listener);
 
         // Hide when clicked on links
         this.$root.$on('clicked::link', listener);
     },
-    mounted: clickOutMixin.mounted,
-    destroyed: clickOutMixin.destroyed,
-    watch: {
-        visible(state, old) {
-            if (state === old) {
-                return; // Avoid duplicated emits
-            }
-
-            if (state) {
-                this.$root.$emit('shown::dropdown', this);
-                /*
-                 If this is a touch-enabled device we add extra
-                 empty mouseover listeners to the body's immediate children;
-                 only needed because of broken event delegation on iOS
-                 https://www.quirksmode.org/blog/archives/2014/02/mouse_event_bub.html
-                 */
-                if (typeof document !== 'undefined' && 'ontouchstart' in document.documentElement) {
-                    const children = Array.prototype.slice.call(document.body.children);
-                    children.forEach(el => {
-                        el.addEventListener('mouseover', this.noop);
-                    });
-                }
-            } else {
-                this.$root.$emit('hidden::dropdown', this);
-                /*
-                 If this is a touch-enabled device we remove the extra
-                 empty mouseover listeners we added for iOS support
-                 */
-                if (typeof document !== 'undefined' && 'ontouchstart' in document.documentElement) {
-                    const children = Array.prototype.slice.call(document.body.children);
-                    children.forEach(el => {
-                        el.removeEventListener('mouseover', this.noop);
-                    });
-                }
-            }
-        }
+    mounted() {
+        this.triggersMounted([this.split ? this.$refs.toggle : this.$refs.button, this.$refs.menu]);
+    },
+    beforeDestroy() {
+        this.triggersBeforeDestroy();
     },
     methods: {
-        ...clickOutMixin.methods,
         noop() {
             // Do nothing event handler (used in visible watch)
         },
         clickOutListener() {
-            this.visible = false;
-        },
-        click(e) {
-            if (this.disabled) {
-                this.visible = false;
-                return;
-            }
-            if (this.split) {
-                this.$emit('click', e);
-                this.$root.$emit('shown::dropdown', this);
-            } else {
-                this.toggle();
+            if (this.classState) {
+                this.toggle(false)
             }
         },
-        toggle() {
-            if (this.disabled) {
-                this.visible = false;
-                return;
-            }
-            this.visible = !this.visible;
-            if (this.visible) {
-                this.$nextTick(function () {
-                    // Clear any items that may have active state left
-                    this.clearItems();
-                    // Focus first visible non-disabled item
-                    const item = this.getFirstItem();
-                    if (item) {
-                        this.focusItem(0, [item]);
-                    }
+        hideComponent() {
+            this.classState = false
+            clearTimeout(this.$data._timeout);
+
+            /*
+             If this is a touch-enabled device we remove the extra
+             empty mouseover listeners we added for iOS support
+             */
+            if (typeof document !== 'undefined' && 'ontouchstart' in document.documentElement) {
+                const children = Array.prototype.slice.call(document.body.children);
+                children.forEach(el => {
+                    el.removeEventListener('mouseover', this.noop);
                 });
             }
         },
+        showComponent() {
+            clearTimeout(this.$data._timeout);
+            if (this.disabled) {
+                this.classState = false;
+                return;
+            }
+            this.classState = true
+
+            /*
+             If this is a touch-enabled device we add extra
+             empty mouseover listeners to the body's immediate children;
+             only needed because of broken event delegation on iOS
+             https://www.quirksmode.org/blog/archives/2014/02/mouse_event_bub.html
+             */
+            if (typeof document !== 'undefined' && 'ontouchstart' in document.documentElement) {
+                const children = Array.prototype.slice.call(document.body.children);
+                children.forEach(el => {
+                    el.addEventListener('mouseover', this.noop);
+                });
+            }
+
+            this.$nextTick(function () {
+                // Clear any items that may have active state left
+                this.clearItems();
+                // Focus first visible non-disabled item
+                const item = this.getFirstItem();
+                if (item) {
+                    this.focusItem(0, [item]);
+                }
+            });
+        },
         onTab() {
-            if (this.visible) {
-                this.visible = false;
+            if (this.classState) {
+                this.classState = false;
             }
         },
         onEsc(e) {
-            if (this.visible) {
-                this.visible = false;
+            if (this.classState) {
+                this.classState = false;
                 e.preventDefault();
                 e.stopPropagation();
                 this.$nextTick(function () {
@@ -160,7 +156,7 @@ export default {
             }
         },
         focusNext(e, up) {
-            if (!this.visible) {
+            if (!this.classState) {
                 return;
             }
             e.preventDefault();
@@ -183,7 +179,7 @@ export default {
             });
         },
         focusHovered(e) {
-            if (!this.visible) {
+            if (!this.classState) {
                 return;
             }
             this.$nextTick(() => {
