+++ lib/mixins/index.js
@@ -7,5 +7,6 @@ import formOptionsMixin from './form-options';
 import linkMixin from './link';
 import listenOnRootMixin from './listen-on-root';
 import popoverMixin from './popover';
+import triggersMixin from './triggers';
 
-export {clickOutMixin, dropdownMixin, formMixin, formCheckBoxMixin, formCustomMixin, formOptionsMixin, linkMixin, listenOnRootMixin, popoverMixin};
+export {clickOutMixin, dropdownMixin, formMixin, formCheckBoxMixin, formCustomMixin, formOptionsMixin, linkMixin, listenOnRootMixin, popoverMixin, triggersMixin};
